+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 
 ### Bug Fixes
 
+- fix circular import error #825 @williballenthin
+
 ### capa explorer IDA Pro plugin
 
 ### Development